RepositoryConfiguration: Add BitBake recipes
BitBake does not have a meta-data file with the name, version, dependencies, etc.
of the project. Therefore we need to provide the name(s) of the recipe(s) that
should be analyzed.

This change adds an optional item to the repository configuration file that maps
paths inside the repository to a list of BitBake recipes for each path.