Analyzer: Construct the managers a bit earlier
Previously there the construction of managers and the mapping of the
definition files was done in one code block. Moving the responsibility
of the construction to the block directly above seems better as it
allows for further readability improvements and as the factory instances
aren't needed anyway by the code below.

Signed-off-by: Frank Viernau <frank.viernau@here.com>